build : auth api login

+++ hono-auth/src/routes/otpEmail.ts
@@ -5,20 +5,23 @@ import { Resend } from "resend";
 import { generateOTP, hashOTP } from "../utils/generateOtp";
 import { eq } from "drizzle-orm";
 import { auth } from "..";
+import { internalPassword } from "../utils/crypto";
 
 const otpRoute = new Hono();
-
-const resend = new Resend(process.env.RESEND_API!);
-
+if (!process.env.RESEND_API) {
+  throw new Error("RESEND_API is missing");
+}
+if (!process.env.OTP_SECRET) {
+  throw new Error("OTP_SECRET is missing");
+}
+const resend = new Resend(process.env.RESEND_API);
 // ================= SEND OTP =================
 otpRoute.post("/send-otp", async (c) => {
-  console.log("running");
   const { email } = await c.req.json();
 
   const otp = generateOTP();
   const otpHash = hashOTP(otp);
   const expiresAt = new Date(Date.now() + 5 * 60 * 1000);
-  console.log("otp :", otp);
   await db.delete(schema.emailOtps).where(eq(schema.emailOtps.email, email));
 
   await db.insert(schema.emailOtps).values({
@@ -28,17 +31,17 @@ otpRoute.post("/send-otp", async (c) => {
   });
 
   await resend.emails.create({
-    from: "DikaPay <no-reply@yourdomain.com>",
+    from: "DikaPay <no-reply@dikapay.app>",
     to: email,
     subject: "Your OTP Code",
     html: `<h2>Your OTP is ${otp}</h2>
-           <p>This code expires in 5 minutes.</p>`,
+           <p>This code expires in 3 minutes.</p>`,
   });
-
   return c.json({ success: true });
 });
 
 // ================= VERIFY OTP =================
+
 otpRoute.post("/verify-otp", async (c) => {
   const { email, otp } = await c.req.json();
   let session;
@@ -74,28 +77,25 @@ otpRoute.post("/verify-otp", async (c) => {
   }
 
   await db.delete(schema.emailOtps).where(eq(schema.emailOtps.email, email));
-
-  const randomPassword = crypto.randomUUID();
-
   try {
-    await auth.api.signUpEmail({
+    session = await auth.api.signUpEmail({
       body: {
         name: email,
         email,
-        password: randomPassword,
+        password: internalPassword(email),
       },
     });
   } catch (err) {
-    // user อาจมีอยู่แล้ว
+    session = await auth.api.signInEmail({
+      body: {
+        email,
+        password: internalPassword(email),
+      },
+    });
+    console.log("session", session);
   }
 
-  session = await auth.api.signInEmail({
-    body: {
-      email,
-      password: randomPassword,
-    },
-  });
-  return c.json({ success: session });
+  return c.json({ token: session.token, user: session.user });
 });
 
 export default otpRoute;
